feat!: use OpenSSL instead of node-forge for DER-to-PEM conversion (#4)
This commit removes node-forge as a dependency and the conversion
logic for X.509 certificates from DER to PEM format that we used it for,
and replaces that logic with direct calls to OpenSSL functions
(which are always available in Node.js).

The main downside of this approach is that it removes the ABI stability
guarantees we get from Node-API and instead makes us dependent on the
OpenSSL ABI as well. Currently, the functions used in this code are
compatible with OpenSSL 1.1.x and 3.x, which should be good enough
in the present and for at least a good number of years into the future.

The main upside of this approach is that it is significantly faster,
and should enable applications that use this package to get the
system certificate list to do so by default, rather than as an opt-in
feature.